big refactor; using `binrw` to parse `.cur` now
the reason i decided this was that `zerocopy`'s philosophy of, well,
not copying, just didn't have much importance in this project.

`.cur` files are kilobytes. messing with lifetimes and references
just isn't worth it for that miniscule return

(also i'm gonna need to copy anyway to convert to xcursor)